+++ app/src/main/java/com/example/gymtracker/data/model/RecipeWithIngredients.kt
@@ -31,5 +31,5 @@ data class RecipeWithDetails(
 
 data class IngredientDetails(
     val foodTemplate: FoodTemplate,
-    val grams: Int
+    val grams: Float
 )
